pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sgmriqa"
version = "1.0.0"
description = "SGMRI-QA: Multi-Frame Spatially Grounded Reasoning Across Volumetric MRI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Lama Moukheiber"},
]
dependencies = [
    "Pillow>=10.0.0",
    "tqdm>=4.60.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "google-generativeai>=0.5.0",
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "bert-score>=0.3.13",
    "nltk>=3.8.0",
    "rouge-score>=0.1.2",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
vllm = ["vllm>=0.6.0"]
anthropic = ["anthropic>=0.18.0"]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
